Remove 32-bit x86 support entirely
