+++ litebox_shim_linux/src/syscalls/misc.rs
@@ -62,8 +62,6 @@ const SYS_INFO: litebox_common_linux::Utsname = litebox_common_linux::Utsname {
     version: to_fixed_size_array::<65>("5.11.0"),
     #[cfg(target_arch = "x86_64")]
     machine: to_fixed_size_array::<65>("x86_64"),
-    #[cfg(target_arch = "x86")]
-    machine: to_fixed_size_array::<65>("x86"),
     domainname: to_fixed_size_array::<65>(""),
 };
 
@@ -88,8 +86,6 @@ impl<FS: ShimFS> Task<FS> {
             loads: [0; 3],
             #[cfg(target_arch = "x86_64")]
             totalram: 4 * 1024 * 1024 * 1024,
-            #[cfg(target_arch = "x86")]
-            totalram: 3 * 1024 * 1024 * 1024,
             freeram: 2 * 1024 * 1024 * 1024,
             sharedram: 0, // We don't support shared memory
             bufferram: 0,
