+++ litebox_shim_linux/src/syscalls/net.rs
@@ -1677,179 +1677,6 @@ impl<FS: ShimFS> Task<FS> {
     }
 }
 
-#[cfg(target_arch = "x86")]
-impl<FS: ShimFS> Task<FS> {
-    pub(crate) fn sys_socketcall(&self, call: i32, args: ConstPtr<usize>) -> Result<usize, Errno> {
-        use crate::ToSyscallResult;
-        use litebox_common_linux::SocketcallType;
-        macro_rules! parse_socketcall_args {
-            ($nargs:literal => $func:ident { $($field:ident: $tt:tt),* $(,)? }) => {{
-                let args = args.to_owned_slice($nargs).ok_or(Errno::EFAULT)?;
-                self.$func (
-                    $(parse_socketcall_args!(@convert args $tt )),*
-                ).to_syscall_result()
-            }};
-
-            // Convert with pointer marker - use from_usize for pointer types
-            (@convert $args:ident [ $idx:literal ]) => {
-                <_ as litebox_common_linux::ReinterpretUsizeAsPtr<_>>::reinterpret_usize_as_ptr($args[$idx])
-            };
-
-            // Convert without marker - use reinterpret_truncated_from_usize for regular types
-            (@convert $args:ident $idx:literal) => {
-                litebox_common_linux::ReinterpretTruncatedFromUsize::reinterpret_truncated_from_usize($args[$idx])
-            };
-
-            (@convert $args:ident $v:ident) => {
-                $v
-            };
-        }
-
-        let socketcall_type = SocketcallType::try_from(call).map_err(|_| Errno::EINVAL)?;
-        match socketcall_type {
-            SocketcallType::Socket => {
-                parse_socketcall_args!(3 => sys_socket {
-                    domain: 0,
-                    type_and_flags: 1,
-                    protocol: 2,
-                })
-            }
-            SocketcallType::Socketpair => {
-                parse_socketcall_args!(4 => sys_socketpair {
-                    domain: 0,
-                    type_and_flags: 1,
-                    protocol: 2,
-                    sv: [ 3 ],
-                })
-            }
-            SocketcallType::Bind => {
-                parse_socketcall_args!(3 => sys_bind {
-                    sockfd: 0,
-                    sockaddr: [ 1 ],
-                    addrlen: 2,
-                })
-            }
-            SocketcallType::Connect => {
-                parse_socketcall_args!(3 => sys_connect {
-                    sockfd: 0,
-                    sockaddr: [ 1 ],
-                    addrlen: 2,
-                })
-            }
-            SocketcallType::Listen => {
-                parse_socketcall_args!(2 => sys_listen {
-                    sockfd: 0,
-                    backlog: 1,
-                })
-            }
-            SocketcallType::Accept => {
-                let flags = SockFlags::empty();
-                parse_socketcall_args!(3 => sys_accept {
-                    sockfd: 0,
-                    addr: [ 1 ],
-                    addrlen: [ 2 ],
-                    flags: flags,
-                })
-            }
-            SocketcallType::Accept4 => {
-                parse_socketcall_args!(4 => sys_accept {
-                    sockfd: 0,
-                    addr: [ 1 ],
-                    addrlen: [ 2 ],
-                    flags: 3,
-                })
-            }
-            SocketcallType::Send => {
-                let addr = None;
-                let addrlen = 0;
-                parse_socketcall_args!(4 => sys_sendto {
-                    sockfd: 0,
-                    buf: [ 1 ],
-                    len: 2,
-                    flags: 3,
-                    addr: addr,
-                    addrlen: addrlen,
-                })
-            }
-            SocketcallType::Sendto => {
-                parse_socketcall_args!(6 => sys_sendto {
-                    sockfd: 0,
-                    buf: [ 1 ],
-                    len: 2,
-                    flags: 3,
-                    addr: [ 4 ],
-                    addrlen: 5,
-                })
-            }
-            SocketcallType::Recv => {
-                let addr = None;
-                let addrlen = MutPtr::from_usize(0);
-                parse_socketcall_args!(4 => sys_recvfrom {
-                    sockfd: 0,
-                    buf: [ 1 ],
-                    len: 2,
-                    flags: 3,
-                    addr: addr,
-                    addrlen: addrlen,
-                })
-            }
-            SocketcallType::Recvfrom => {
-                parse_socketcall_args!(6 => sys_recvfrom {
-                    sockfd: 0,
-                    buf: [ 1 ],
-                    len: 2,
-                    flags: 3,
-                    addr: [ 4 ],
-                    addrlen: [ 5 ],
-                })
-            }
-            SocketcallType::GetSockname => {
-                parse_socketcall_args!(3 => sys_getsockname {
-                    sockfd: 0,
-                    addr: [ 1 ],
-                    addrlen: [ 2 ],
-                })
-            }
-            SocketcallType::GetPeername => {
-                parse_socketcall_args!(3 => sys_getpeername {
-                    sockfd: 0,
-                    addr: [ 1 ],
-                    addrlen: [ 2 ],
-                })
-            }
-            SocketcallType::Setsockopt => {
-                parse_socketcall_args!(5 => sys_setsockopt {
-                    sockfd: 0,
-                    level: 1,
-                    optname: 2,
-                    optval: [ 3 ],
-                    optlen: 4,
-                })
-            }
-            SocketcallType::Getsockopt => {
-                parse_socketcall_args!(5 => sys_getsockopt {
-                    sockfd: 0,
-                    level: 1,
-                    optname: 2,
-                    optval: [ 3 ],
-                    optlen: [ 4 ],
-                })
-            }
-            SocketcallType::Sendmsg => {
-                parse_socketcall_args!(3 => sys_sendmsg {
-                    sockfd: 0,
-                    msg: [ 1 ],
-                    flags: 2,
-                })
-            }
-            _ => {
-                log_unsupported!("socketcall type {socketcall_type:?} is not supported");
-                Err(Errno::EINVAL)
-            }
-        }
-    }
-}
-
 #[cfg(target_os = "linux")]
 #[cfg(test)]
 mod tests {
